fix(ui): small authentication fixes and adjustments
- fix nextauth error caused by wrong auth/session path
- allow enabling authentication without https
- minor refactor to match project structure
- prettify visual components
- remove setting access_token cookie
