--- apps/beeai-ui/src/app/(signin)/signin/signin.module.scss
@@ -1,55 +1,0 @@
-/**
- * Copyright 2025 © BeeAI a Series of LF Projects, LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-.bannerBackground {
-  background-image: url('/BeeAI-Banner.svg');
-  background-size: cover;
-  background-position: 50% 50%;
-  background-repeat: no-repeat no-repeat;
-  resize: both;
-  inline-size: 100%;
-  margin-block-end: 16px;
-  margin-inline-start: 0px;
-  block-size: 100%;
-}
-
-.buttonList {
-  margin: 16px;
-  position: relative;
-  align-content: center;
-  inset-inline-start: 1rem;
-}
-
-.buttonEntry {
-  block-size: 48px;
-}
-
-.beeButton {
-  margin-inline-end: 16px;
-}
-
-.loginWrapper {
-  background-color: var(--cds-layer-01, #f4f4f4);
-  box-sizing: content-box;
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  padding-inline-start: 4rem;
-  padding-inline-end: 6rem;
-  padding-block-start: 8.875rem;
-  inline-size: 31.25rem;
-  block-size: 100%;
-}
-
-.loginMain {
-  font-size: 2rem;
-  .bolded {
-    font-weight: 600;
-  }
-}
-
-.beeaiLogin {
-  margin: 16px;
-}
