--- apps/beeai-ui/src/components/Avatar/Avatar.module.scss
@@ -1,13 +1,0 @@
-/**
- * Copyright 2025 © BeeAI a Series of LF Projects, LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-.avatar {
-  position: absolute;
-  inset-inline-end: 80px;
-}
-
-.hidden {
-  display: none;
-}
