+++ apps/beeai-ui/src/components/UserNav/UserNav.module.scss
@@ -3,6 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-.button {
+button.button {
   border: 1px solid $border-subtle;
+
+  &.avatar {
+    border-radius: 50%;
+  }
+}
+
+.menu {
+  transform: translate(0, -$spacing-02);
 }
