+++ apps/beeai-ui/src/components/layouts/SideNav.module.scss
@@ -12,5 +12,4 @@
 
 .footer {
   padding-inline: $spacing-05;
-  display: flex;
 }
