+++ apps/beeai-ui/src/middleware.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { auth as middleware } from '#auth.ts';
+export { auth as middleware } from '#app/(auth)/auth.ts';
 
 // Read more: https://nextjs.org/docs/app/building-your-application/routing/middleware#matcher
 export const config = {
