+++ apps/beeai-ui/src/utils/constants.ts
@@ -35,6 +35,6 @@ export const COMPOSE_LINK = `${DOCUMENTATION_LINK}/how-to/compose-agents`;
 
 export const LF_PROJECTS_LINK = 'https://lfprojects.org/';
 
-export const OIDC_ENABLED = process.env.NEXT_PUBLIC_OIDC_ENABLED === 'true';
+export const OIDC_ENABLED = process.env.OIDC_ENABLED === 'true';
 
-export const AUTH_BASEPATH = '/auth';
+// export const AUTH_BASEPATH = '/api/auth';
